requirements.txt
twitter-scraper-selenium==6.1.2
tweepy==4.14.0
requests==2.27.1
twitter-api-client==0.10.22